style changes, disable to delete category if not empty.

+++ Control/ItemListControl.php
@@ -19,25 +19,41 @@ class ItemListControl extends Control {
 	public function initialize() {
 		global $loggedIn;
 		global $admin;
-
-		if($loggedIn && $admin) {
-			if(isset($_POST['categoryName'])) {
+		$this->model->fail = FALSE;
+		if(isset($_POST['categoryName'])) {
+			if($loggedIn && $admin) {
 				$categoryName = sanitizeString($_POST['categoryName']);
 				$this->addCategory($categoryName);
+			} else {
+				$this->model->error = "<span class='error'>You're not logged in, or must be admin to add/edit.</span>";
 			}
-			if(isset($_POST['deleteCategory'])) {
+		}
+		if(isset($_POST['deleteCategory'])) {
+			if($loggedIn && $admin) {
 				$delete = (int)sanitizeString($_POST['categoryDelete']);
 				$category = CategoryDAO::selectById($delete);
-				CategoryDAO::delete($category);
+				$items = ItemDAO::selectByCategoryId($delete);
+				if($items!=NULL){
+					$this->model->fail = TRUE;
+					$this->model->error = "<span class='error'>Can't delete category with items in it.</span>";
+				}else{
+					CategoryDAO::delete($category);
+				}
+
+			} else {
+				$this->model->error = "<span class='error'>You're not logged in, or must be admin to add/edit.</span>";
 			}
-			if(isset($_POST['DeleteItem'])) {
+		}
+		if(isset($_POST['DeleteItem'])) {
+			if($loggedIn && $admin) {
 				$itemDeleteId = (int)sanitizeString($_POST['itemId']);
 				$item = ItemDAO::selectById($itemDeleteId);
 				ItemDAO::delete($item);
+			} else {
+				$this->model->error = "<span class='error'>You're not logged in, or must be admin to add/edit.</span>";
 			}
-		} else {
-			$this->model->error = "<span class='error'>You're not logged in, or must be admin to add/edit.</span>";
 		}
+
 		$this->loadCategories();
 		$n = count($this->categories);
 		for ($i = 0; $i < $n; $i++) {
